Temperature Converter
This app is able to convert the temperature.
The best app for school, college and work! If you are a student, it will helps you to learn temperature calculations.

Note: On the Fahrenheit scale, the freezing point of water is 32 degrees Fahrenheit (°F) and the boiling point 212 °F (at standard atmospheric pressure). 
On the Celsius scale, the freezing point of water is 0 *C and the boiling point 100 °C.